backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
fastapi-cors==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
stripe==7.4.0
httpx==0.25.1
anthropic==0.7.11
pandas==2.1.3
python-multipart==0.0.6
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
gunicorn==21.2.0
